Move coverage tests to their natural homes
AI disclosure: I used OpenAI Codex (GPT-5.6) to apply the requested test organization changes, review the diff, and verify formatting and the five affected tests.
