+++ src/icalendar/tests/prop/test_unit.py
@@ -344,6 +344,21 @@ class TestProp(unittest.TestCase):
         cats = list(c)
         assert cats == ["APPOINTMENT", "EDUCATION"]
 
+    def test_vcategory_hash(self):
+        from icalendar.prop import vCategory
+
+        category = vCategory(["WORK", "PERSONAL"])
+        assert hash(category) == hash(tuple(category.cats))
+
+    def test_vcategory_repr(self):
+        from icalendar.prop import vCategory
+
+        category = vCategory(["WORK", "PERSONAL"])
+        category_repr = repr(category)
+        assert category_repr == (
+            "vCategory([vText(b'WORK'), vText(b'PERSONAL')], params=Parameters({}))"
+        )
+
     def test_prop_TypesFactory(self):
         from icalendar.prop import TypesFactory
